+++ tests/test_utils/test_fab_cmd_mkdir_utils.py
@@ -18,7 +18,6 @@ from fabric_cli.utils.fab_cmd_mkdir_utils import (
     get_connection_config_from_params,
     get_params_per_item_type,
 )
-from fabric_cli.utils.fab_util import is_valid_guid, is_valid_iso8601_timestamp
 
 
 def test_fabric_data_pipelines_workspace_identity_no_params_success():
@@ -222,82 +221,6 @@ class TestFindMpeConnection:
             )
             assert "privateEndpointConnections" in called_url
             assert "api-version=2023-11-01" in called_url
-
-
-class TestIsValidGuid:
-    """Test cases for is_valid_guid function."""
-
-    def test_valid_guid_lowercase(self):
-        """Test valid GUID with lowercase letters."""
-        assert is_valid_guid("12345678-1234-1234-1234-123456789abc")
-
-    def test_valid_guid_uppercase(self):
-        """Test valid GUID with uppercase letters."""
-        assert is_valid_guid("12345678-1234-1234-1234-123456789ABC")
-
-    def test_valid_guid_mixed_case(self):
-        """Test valid GUID with mixed case letters."""
-        assert is_valid_guid("12345678-abcd-ABCD-1234-123456789AbC")
-
-    def test_invalid_guid_wrong_format(self):
-        """Test invalid GUID with wrong format."""
-        assert not is_valid_guid("not-a-guid")
-
-    def test_invalid_guid_missing_hyphens(self):
-        """Test invalid GUID with missing hyphens."""
-        assert not is_valid_guid("1234567812341234123412345678abc")
-
-    def test_invalid_guid_extra_characters(self):
-        """Test invalid GUID with extra characters."""
-        assert not is_valid_guid("12345678-1234-1234-1234-123456789abcd")
-
-    def test_invalid_guid_empty_string(self):
-        """Test empty string is not a valid GUID."""
-        assert not is_valid_guid("")
-
-
-class TestIsValidIso8601Timestamp:
-    """Test cases for is_valid_iso8601_timestamp function."""
-
-    def test_valid_timestamp_utc_z_suffix(self):
-        """Test valid ISO 8601 timestamp with Z suffix."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00Z")
-
-    def test_valid_timestamp_with_milliseconds(self):
-        """Test valid ISO 8601 timestamp with milliseconds."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00.123Z")
-
-    def test_valid_timestamp_with_microseconds(self):
-        """Test valid ISO 8601 timestamp with microseconds."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00.123456Z")
-
-    def test_valid_timestamp_positive_offset(self):
-        """Test valid ISO 8601 timestamp with positive timezone offset."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00+05:30")
-
-    def test_valid_timestamp_negative_offset(self):
-        """Test valid ISO 8601 timestamp with negative timezone offset."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00-08:00")
-
-    def test_valid_timestamp_zero_offset(self):
-        """Test valid ISO 8601 timestamp with zero offset."""
-        assert is_valid_iso8601_timestamp("2024-01-15T10:30:00+00:00")
-
-    def test_invalid_timestamp_no_timezone(self):
-        """Test timestamp without timezone is invalid."""
-        assert not is_valid_iso8601_timestamp("2024-01-15T10:30:00")
-
-    def test_invalid_timestamp_date_only(self):
-        """Test date-only string is invalid."""
-        assert not is_valid_iso8601_timestamp("2024-01-15")
-
-    def test_invalid_timestamp_wrong_format(self):
-        """Test invalid timestamp format."""
-        assert not is_valid_iso8601_timestamp("not-a-timestamp")
-
-    def test_invalid_timestamp_empty_string(self):
-        """Test empty string is not a valid timestamp."""
-        assert not is_valid_iso8601_timestamp("")
 
 
 class TestSQLDatabaseRestore:
